# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for monitoring Keycloak instances, managing alerts, and security operations",
        "title": "Keycloak Monitoring Tool API",
        "termsOfService": "https://defensepoint.com/terms",
        "contact": {
            "name": "DefensePoint Support",
            "url": "https://defensepoint.com/support",
            "email": "support@defensepoint.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/rbac/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current user's roles, permissions, and tenant policies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rbac"
                ],
                "summary": "Get current user RBAC info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by tenant ID",
                        "name": "tenant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "email_verified": {
                                    "type": "boolean"
                                },
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "permissions": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "preferred_username": {
                                    "type": "string"
                                },
                                "roles": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_domain.UserRole"
                                    }
                                },
                                "subject": {
                                    "type": "string"
                                },
                                "tenant_policies": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_domain.TenantPolicy"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/rbac/permissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of all available permissions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rbac"
                ],
                "summary": "List all permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by resource name",
                        "name": "resource",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "permissions": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_domain.Permission"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/rbac/policies/{policyID}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a tenant access policy",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rbac"
                ],
                "summary": "Delete tenant policy",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Policy ID",
                        "name": "policyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid policy ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/rbac/users/{userID}/roles/{roleID}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a role assignment from a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rbac"
                ],
                "summary": "Remove role from user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "roleID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tenant ID (for tenant-scoped roles)",
                        "name": "tenant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user or role ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_DefensePoint_keycloak-monitoring_internal_domain.UserRole": {
            "type": "object",
            "properties": {
                "assigned_at": {
                    "type": "string"
                },
                "assigned_by": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "$ref": "#/definitions/github_com_DefensePoint_keycloak-monitoring_internal_domain.Role"
                },
                "role_id": {
                    "type": "integer"
                },
                "tenant_id": {
                    "description": "Null = global, non-null = tenant-scoped",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}